requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
requests
sentence-transformers
faiss-cpu
numpy
pandas
nltk
scikit-learn
beautifulsoup4
lxml
pydantic
emoji
selenium
webdriver_manager
langdetect
bertopic
umap-learn
hdbscan
pymupdf 
pymongo 
tiktoken 
